+++ x/tx/signing/aminojson/json_marshal.go
@@ -37,13 +37,14 @@ type EncoderOptions struct {
 // Encoder is a JSON encoder that uses the Amino JSON encoding rules for protobuf messages.
 type Encoder struct {
 	// maps cosmos_proto.scalar -> field encoder
-	scalarEncoders  map[string]FieldEncoder
-	messageEncoders map[string]MessageEncoder
-	fieldEncoders   map[string]FieldEncoder
-	fileResolver    signing.ProtoFileResolver
-	typeResolver    protoregistry.MessageTypeResolver
-	doNotSortFields bool
-	indent          string
+	cosmosProtoScalarEncoders map[string]FieldEncoder
+	aminoMessageEncoders      map[string]MessageEncoder
+	aminoFieldEncoders        map[string]FieldEncoder
+	protoTypeEncoders         map[string]MessageEncoder
+	fileResolver              signing.ProtoFileResolver
+	typeResolver              protoregistry.MessageTypeResolver
+	doNotSortFields           bool
+	indent                    string
 }
 
 // NewEncoder returns a new Encoder capable of serializing protobuf messages to JSON using the Amino JSON encoding
@@ -56,17 +57,22 @@ func NewEncoder(options EncoderOptions) Encoder {
 		options.TypeResolver = protoregistry.GlobalTypes
 	}
 	enc := Encoder{
-		scalarEncoders: map[string]FieldEncoder{
+		cosmosProtoScalarEncoders: map[string]FieldEncoder{
 			"cosmos.Dec": cosmosDecEncoder,
 			"cosmos.Int": cosmosIntEncoder,
 		},
-		messageEncoders: map[string]MessageEncoder{
+		aminoMessageEncoders: map[string]MessageEncoder{
 			"key_field":        keyFieldEncoder,
 			"module_account":   moduleAccountEncoder,
 			"threshold_string": thresholdStringEncoder,
 		},
-		fieldEncoders: map[string]FieldEncoder{
+		aminoFieldEncoders: map[string]FieldEncoder{
 			"legacy_coins": nullSliceAsEmptyEncoder,
+		},
+		protoTypeEncoders: map[string]MessageEncoder{
+			"google.protobuf.Timestamp": marshalTimestamp,
+			"google.protobuf.Duration":  marshalDuration,
+			"google.protobuf.Any":       marshalAny,
 		},
 		fileResolver:    options.FileResolver,
 		typeResolver:    options.TypeResolver,
@@ -86,10 +92,10 @@ func NewEncoder(options EncoderOptions) Encoder {
 //	  ...
 //	}
 func (enc Encoder) DefineMessageEncoding(name string, encoder MessageEncoder) Encoder {
-	if enc.messageEncoders == nil {
-		enc.messageEncoders = map[string]MessageEncoder{}
+	if enc.aminoMessageEncoders == nil {
+		enc.aminoMessageEncoders = map[string]MessageEncoder{}
 	}
-	enc.messageEncoders[name] = encoder
+	enc.aminoMessageEncoders[name] = encoder
 	return enc
 }
 
@@ -107,10 +113,10 @@ func (enc Encoder) DefineMessageEncoding(name string, encoder MessageEncoder) En
 //	  ...
 //	}
 func (enc Encoder) DefineFieldEncoding(name string, encoder FieldEncoder) Encoder {
-	if enc.fieldEncoders == nil {
-		enc.fieldEncoders = map[string]FieldEncoder{}
+	if enc.aminoFieldEncoders == nil {
+		enc.aminoFieldEncoders = map[string]FieldEncoder{}
 	}
-	enc.fieldEncoders[name] = encoder
+	enc.aminoFieldEncoders[name] = encoder
 	return enc
 }
 
@@ -123,10 +129,27 @@ func (enc Encoder) DefineFieldEncoding(name string, encoder FieldEncoder) Encode
 //	  ...
 //	}
 func (enc Encoder) DefineScalarEncoding(name string, encoder FieldEncoder) Encoder {
-	if enc.scalarEncoders == nil {
-		enc.scalarEncoders = map[string]FieldEncoder{}
+	if enc.cosmosProtoScalarEncoders == nil {
+		enc.cosmosProtoScalarEncoders = map[string]FieldEncoder{}
 	}
-	enc.scalarEncoders[name] = encoder
+	enc.cosmosProtoScalarEncoders[name] = encoder
+	return enc
+}
+
+// DefineTypeEncoding defines a custom encoding for a protobuf message type.  The `typeURL` field must match the
+// type of the protobuf message as in the following example. This encoding will be used instead of the default
+// encoding for all usages of the tagged message.
+//
+//	message Foo {
+//	  google.protobuf.Duration type_url = 1;
+//	  ...
+//	}
+
+func (enc Encoder) DefineTypeEncoding(typeURL string, encoder MessageEncoder) Encoder {
+	if enc.protoTypeEncoders == nil {
+		enc.protoTypeEncoders = map[string]MessageEncoder{}
+	}
+	enc.protoTypeEncoders[typeURL] = encoder
 	return enc
 }
 
@@ -209,14 +232,9 @@ func (enc Encoder) marshalMessage(msg protoreflect.Message, writer io.Writer) er
 		return errors.New("nil message")
 	}
 
-	switch msg.Descriptor().FullName() {
-	case timestampFullName:
-		// replicate https://github.com/tendermint/go-amino/blob/8e779b71f40d175cd1302d3cd41a75b005225a7a/json-encode.go#L45-L51
-		return marshalTimestamp(msg, writer)
-	case durationFullName:
-		return marshalDuration(msg, writer)
-	case anyFullName:
-		return enc.marshalAny(msg, writer)
+	// check if we have a custom type encoder for this type
+	if typeEnc, ok := enc.protoTypeEncoders[string(msg.Descriptor().FullName())]; ok {
+		return typeEnc(&enc, msg, writer)
 	}
 
 	if encoder := enc.getMessageEncoder(msg); encoder != nil {
@@ -370,9 +388,3 @@ func (enc Encoder) marshalList(list protoreflect.List, writer io.Writer) error {
 	_, err = io.WriteString(writer, "]")
 	return err
 }
-
-const (
-	timestampFullName protoreflect.FullName = "google.protobuf.Timestamp"
-	durationFullName  protoreflect.FullName = "google.protobuf.Duration"
-	anyFullName       protoreflect.FullName = "google.protobuf.Any"
-)